+++ lib/camerata/secrets.rb
@@ -12,7 +12,7 @@ module Camerata
         MC_USER
         MC_PW
         PRESERVICA_CREDENTIALS
-        RAILS_MASTER_KEY
+        RAILS_MAIN_KEY
         DYNATRACE_TOKEN
         DEPLOY_KEY_ID
         DEPLOY_ACCESS_KEY